Reorganize once more
