+++ packages/hdwallet-provider/src/constructorTypes/Constructor.ts
@@ -32,21 +32,3 @@ export interface CommonOptions {
 }
 
 export type Options = SigningAuthority & CommonOptions;
-
-// extract the mnemonic if that's the style used, or return undefined
-export const getMnemonic = (
-  signingAuthority: SigningAuthority
-): Mnemonic | undefined => {
-  if ("mnemonic" in signingAuthority) {
-    return signingAuthority.mnemonic;
-  }
-};
-
-// extract the private keys if that's the style used, or return undefined
-export const getPrivateKeys = (
-  signingAuthority: SigningAuthority
-): PrivateKey[] | undefined => {
-  if ("privateKeys" in signingAuthority) {
-    return signingAuthority.privateKeys;
-  }
-};
