+++ packages/hdwallet-provider/src/getMnemonic.ts
@@ -1,0 +1,11 @@
+import { SigningAuthority } from "./constructorTypes/Constructor";
+import { Mnemonic } from "./constructorTypes/types";
+
+// extract the mnemonic if that's the style used, or return undefined
+export const getMnemonic = (
+  signingAuthority: SigningAuthority
+): Mnemonic | undefined => {
+  if ("mnemonic" in signingAuthority) {
+    return signingAuthority.mnemonic;
+  }
+};
