+++ packages/hdwallet-provider/src/getPrivateKeys.ts
@@ -1,0 +1,11 @@
+import { PrivateKey } from "./constructorTypes/types";
+import { SigningAuthority } from "./constructorTypes/Constructor";
+
+// extract the private keys if that's the style used, or return undefined
+export const getPrivateKeys = (
+  signingAuthority: SigningAuthority
+): PrivateKey[] | undefined => {
+  if ("privateKeys" in signingAuthority) {
+    return signingAuthority.privateKeys;
+  }
+};
